Modify the semantic of "include: " entry in YAML files
Before this commit, the value of the include directive was directly
pasted into the generated file so the YAML file author had to take care
of include compilation paths and make sure the file where found at
compile time.

Now, the absolute path of the YAML file is prepended to the value of the
include directive.

Example: we have a file named /path/to/components/memory/memory.yml that
contains:

component:
[...]
  include: memory.h
[...]

The resulting include directive in the generated memory.yml.h will be:

\#include "/path/to/components/memory/memory.h"

Thus, all YAML files with relative paths to the root of their
repositories in their include directives must be modified, and the
    include_directories("${PROJECT_SOURCE_DIR}")
directive in all root CMakeLists.txt files can be removed.