rabbits: Ignore keys starting with '.' in description
When checking for unused values in the platform description, ignore the
ones that have their root key starting with a dot. For example:

foo:
  bar:
    baz: 4

If foo.bar.baz is not used by Rabbits, it will generate a warning such as
[ warn] Unknown parameter `baz` at xxx.yml:2:9

However if the root key starts with a dot, the whole content is ignored:

.foo:
  bar:
    baz: 3

won't generate the warning.

It is useful in YAML descriptions to create templates that are used as
references, for example:

.my-common-components: &common
  c1:
    type: foo
    bindings:
      ...
  c2:
    type: bar
    bindings:
      ...

platforms:
  p1:
    ...
    components:
      *common

platforms:
  p2:
    ...
    components:
      *common