Add the resource manager.
Resources are misc files that can be added through cmake commands and
retrieved at execution with the ResourceManager class.

Resources are split into inventories, to avoid name clashes.

  / --- inventory1 --- resource0
    |               \_ resource1
    |
    \__ inventory1 --- resource0
                    \_ resource1

To add a resource, use the rabbits_add_resource(source_file inventory dest)
cmake function.

Example:
    rabbits_add_resource("index.html" "raspberry-pi" "board-gui/index.html")

Then, at runtime, use the ConfigManager instance to get the
ResourceManager and your resource.

Example:
    c.get_resource_mananger()
        .get_inventory("raspberry-pi")
        .get_resource("board-gui/index.html")
        .get_absolute_path(); // or get_absolute_uri()

The dest argument of the rabbits_add_resource cmake function is
optional. If not given, the resource name will be the same as the source
file name.