Remove the ConfigManager singleton.
This was the last singleton of librabbits. It was only used by the
logger subsystem, by LOG/LOG_F macros to get the root loggers.

Unfortunately it's hard to get totally ride of this behavior so we keep
a reference onto the root loggers globally. This reference is set by the
ConfigManager at construct time, as early as possible so that the LOG
macros work.