Introduce the ImageLoader class.
This class allows for loading of files and data in various formats. It
has several ImageLoaderHelper classes, each supporting the load of one
format (currently, elf and raw binary).

The ImageLoader is available in the ConfigManager and can be provided
with other ImageLoaderHelper instances to support new formats (through
plugins for example).

On file or data loading, each helper is tested until one is able to load
the payload. The binary loader is always tried in last resort as it
always succeeds since it does not analyse the payload format at all and
load it as-is.