PlatformDescription: keep an hint on datatype when de-serializing
In serialized formats like JSON or YAML, an integer is not the same as a
string ("1" != 1). This is why it's important to keep what kind of data
it was before de-serilization, even though everything is stored as
strings.

This way, when serializing the tree, we can keep the type semantic. For
the moment it is supported for JSON, but not for YAML as the underlying
library (libyaml-cpp) does not give us the type information.